Resolving few review comments, components moving
